requirements.txt
llama-index
llama-index-llms-ollama
llama-index-tools-mcp

fastapi
uvicorn[standard]

# Optional (if needed for async requests or debugging)
httpx
langchain>=0.2.0
langchain-community>=0.0.20
sentence-transformers>=2.2.2
chromadb>=0.4.14
torch>=2.0.0
transformers>=4.37.0
pypdf>=3.0.0
PyPDF2>=3.0.1
aiofiles
python-multipart
jinja2
requests
scikit-learn>=1.4.0

numpy==1.26.4
